Remove duplication and code splitting
See previous commit for details.
