+++ src/components/layout-menu/hooks/use-nudge-display-mode/use-nudge-display-mode.ts
@@ -1,6 +1,7 @@
 import { useGlobal } from 'reactn'
 
-import { nudgeDisplayMode, partiallyApplyNudgeFunction } from '../../utils'
+import { partiallyApplyNudgeFunction } from '../../utils'
+import { DisplayModes } from '../../../../types'
 
 export const useNudgeDisplayMode = (): ((arg0: 'UP' | 'DOWN') => void) => {
   const [activeDisplayMode, setActiveDisplayMode] = useGlobal(
@@ -11,4 +12,19 @@ export const useNudgeDisplayMode = (): ((arg0: 'UP' | 'DOWN') => void) => {
     activeDisplayMode,
     setActiveDisplayMode,
   })
+}
+
+type PartialParams = {
+  readonly activeDisplayMode: DisplayModes
+  readonly setActiveDisplayMode: (arg0: DisplayModes) => void
+}
+
+const nudgeDisplayMode = (partialParams: PartialParams): void => {
+  const { activeDisplayMode, setActiveDisplayMode } = partialParams
+
+  if (activeDisplayMode === DisplayModes.Degree) {
+    setActiveDisplayMode(DisplayModes.Pitch)
+  } else {
+    setActiveDisplayMode(DisplayModes.Degree)
+  }
 }
