+++ src/components/layout-menu/hooks/use-nudge-experience-mode/use-nudge-experience-mode.ts
@@ -1,6 +1,7 @@
 import { useGlobal } from 'reactn'
 
-import { nudgeExperienceMode, partiallyApplyNudgeFunction } from '../../utils'
+import { partiallyApplyNudgeFunction } from '../../utils'
+import { ExperienceModes } from '../../../../types'
 
 export const useNudgeExperienceMode = (): ((arg0: 'UP' | 'DOWN') => void) => {
   const [activeExperienceMode, setActiveExperienceMode] = useGlobal(
@@ -11,4 +12,19 @@ export const useNudgeExperienceMode = (): ((arg0: 'UP' | 'DOWN') => void) => {
     activeExperienceMode,
     setActiveExperienceMode,
   })
+}
+
+type PartialParams = {
+  readonly activeExperienceMode: ExperienceModes
+  readonly setActiveExperienceMode: (arg0: ExperienceModes) => void
+}
+
+const nudgeExperienceMode = (partialParams: PartialParams): void => {
+  const { activeExperienceMode, setActiveExperienceMode } = partialParams
+
+  if (activeExperienceMode === ExperienceModes.Quiz) {
+    setActiveExperienceMode(ExperienceModes.Explore)
+  } else {
+    setActiveExperienceMode(ExperienceModes.Quiz)
+  }
 }
