+++ src/components/layout-menu/hooks/use-nudge-harp-strata-by-apparatus/use-nudge-harp-strata-by-apparatus.ts
@@ -1,9 +1,11 @@
 import { useGlobal } from 'reactn'
+import { getHarpStrata, getApparatusIds } from 'harpstrata'
+import type { HarpStrata, ApparatusIds } from 'harpstrata'
 
-import {
-  nudgeHarpStrataByApparatus,
-  partiallyApplyNudgeFunction,
-} from '../../utils'
+import { partiallyApplyNudgeFunction } from '../../utils'
+import { DisplayModes } from '../../../../types'
+import type { SetActiveHarpStrata } from '../../../../types'
+import { getPropsForHarpStrata } from '../../../../helpers'
 
 export const useNudgeHarpStrataByApparatus = (): ((
   arg0: 'UP' | 'DOWN'
@@ -14,4 +16,53 @@ export const useNudgeHarpStrataByApparatus = (): ((
     activeHarpStrata,
     setActiveHarpStrata,
   })
+}
+
+type PartialParams = {
+  readonly activeHarpStrata: HarpStrata
+  readonly setActiveHarpStrata: SetActiveHarpStrata
+}
+
+const nudgeHarpStrataByApparatus = (
+  partialParams: PartialParams,
+  direction: 'UP' | 'DOWN'
+): void => {
+  const { activeHarpStrata, setActiveHarpStrata } = partialParams
+  const {
+    apparatus: { id: apparatusId },
+  } = activeHarpStrata
+
+  const nextHarpStrataProps = getPropsForHarpStrata(
+    {
+      ...activeHarpStrata,
+      apparatus: {
+        ...activeHarpStrata.apparatus,
+        id: getNextId(apparatusId, direction),
+      },
+    },
+    DisplayModes.Degree
+  )
+
+  setActiveHarpStrata(getHarpStrata(nextHarpStrataProps))
+}
+
+const getNextId = (
+  apparatusId: ApparatusIds,
+  direction: 'UP' | 'DOWN'
+): ApparatusIds => {
+  const apparatusIdList = getApparatusIds()
+  const apparatusIdIndex = apparatusIdList.indexOf(apparatusId)
+
+  if (direction === 'UP') {
+    const [, nextApparatusId] = [
+      ...apparatusIdList.slice(apparatusIdIndex),
+      ...apparatusIdList.slice(0, apparatusIdIndex),
+    ]
+    return nextApparatusId
+  }
+  const [previousApparatusId] = [
+    ...apparatusIdList.slice(apparatusIdIndex),
+    ...apparatusIdList.slice(0, apparatusIdIndex),
+  ].reverse()
+  return previousApparatusId
 }
