+++ src/components/layout-menu/utils/index.ts
@@ -1,4 +1,1 @@
-export { nudgeHarpStrataByApparatus } from './nudge-harp-strata-by-apparatus'
-export { nudgeExperienceMode } from './nudge-experience-mode'
-export { nudgeDisplayMode } from './nudge-display-mode'
 export { partiallyApplyNudgeFunction } from './paritally-apply-nudge-function'
