--- src/components/layout-menu/utils/nudge-display-mode/index.ts
@@ -1,1 +1,0 @@
-export { nudgeDisplayMode } from './nudge-display-mode'
