--- src/components/layout-menu/utils/nudge-experience-mode/index.ts
@@ -1,1 +1,0 @@
-export { nudgeExperienceMode } from './nudge-experience-mode'
