--- src/components/layout-menu/utils/nudge-experience-mode/nudge-experience-mode.ts
@@ -1,16 +1,0 @@
-import { ExperienceModes } from '../../../../types'
-
-type Props = {
-  readonly activeExperienceMode: ExperienceModes
-  readonly setActiveExperienceMode: (arg0: ExperienceModes) => void
-}
-
-export const nudgeExperienceMode = (partialParams: Props): void => {
-  const { activeExperienceMode, setActiveExperienceMode } = partialParams
-
-  if (activeExperienceMode === ExperienceModes.Quiz) {
-    setActiveExperienceMode(ExperienceModes.Explore)
-  } else {
-    setActiveExperienceMode(ExperienceModes.Quiz)
-  }
-}
