--- src/components/layout-menu/utils/nudge-harp-strata-by-apparatus/index.ts
@@ -1,1 +1,0 @@
-export { nudgeHarpStrataByApparatus } from './nudge-harp-strata-by-apparatus'
